requirements.txt
flask
google-generativeai
chromadb
pandas
numpy
python-dotenv